+++ CHANGES.md
@@ -3,7 +3,9 @@
 List changes at major/minor level. Patches should be straightforward.
 
 ## 1.1.0
-- Add coordinates formatting to display nice coordinates.
+- Remove CommonProperties from the `src/layer/layer-group.ts` file and move the `LayerUid` key into
+  `src/layer/propertiy-key.ts` as `LayerUidKey`.
+- Add coordinate formatting to display nice coordinates.
 
 ## v1.0.1
 - Start removing lodash, better import of it.